Use local framework types
