+++ app/views/AnalyticalFramework/FrameworkForm/PrimaryTagging/SectionsEditor/index.tsx
@@ -35,7 +35,7 @@ import {
 import NonFieldError from '#components/NonFieldError';
 import SortableList, { Attributes, Listeners } from '#components/SortableList';
 
-import { Section } from '#types/newAnalyticalFramework';
+import { Section } from '../../../types';
 import styles from './styles.css';
 
 const SECTIONS_LIMIT = 10;
